Create riddle-game-2020.jav
Final Java project from AP Java as a high schooler sophomore. Part of a collaborative "video game" made with 4 others, where the others created their own challenge for the play to win and progress through the story. Pretty rudimentary and primarily a way to take all the programming we learnt and create something that would load out of an IDE.